__init__.py changes, migrating to a new CLI setup
Start using module level dunders for project information.

Using argparse and commands as objects will greatly help readability,
lessen a lot pain adding and changing commands, and provide a nice
entry point for use with a revamped setup.py.

This not only handles the commands (and subcommands), but it also
handles "--help" pages for the commands.

USAGE:
python3 -m bibliopixel